requirements.txt
tensorflow==2.15.0
numpy
pandas
matplotlib
seaborn
pillow
scikit-learn
pyyaml